# pkg/metastore: Reduce allocs in kv.go