This is a follow up PR related to https://github.com/parca-dev/parca/pull/3064 to reduce allocations in `NormalizeWriteRawRequest`.

```
# new
BenchmarkMakeFunctionID-2   	  858595	      1447 ns/op	      96 B/op	       1 allocs/op
# old
BenchmarkMakeFunctionID-12    	  640981	      1804 ns/op	    1064 B/op	      12 allocs/op
```

```
# new
BenchmarkMakeMappingID-12    	 3670863	       323.4 ns/op	      48 B/op	       1 allocs/op
# old
BenchmarkMakeMappingID-12    	 2304825	       531.8 ns/op	     424 B/op	       8 allocs/op
```

```
# new
BenchmarkMakeStacktraceID-12    	  322155	      3702 ns/op	     144 B/op	       1 allocs/op
# old
BenchmarkMakeStacktraceID-12    	  228817	      5012 ns/op	    2864 B/op	      31 allocs/op
```

This helped `NormalizeWriteRawRequest` to shave 94 K allocs, 7.9 MB, and 6.4 ms per operation.

```
$ benchstat bench-old.txt bench-new.txt 
name                         old time/op    new time/op    delta
NormalizeWriteRawRequest-12    43.2ms ± 7%    36.8ms ± 7%  -14.82%  (p=0.000 n=49+46)

name                         old alloc/op   new alloc/op   delta
NormalizeWriteRawRequest-12    52.4MB ± 1%    44.5MB ± 1%  -15.20%  (p=0.000 n=50+50)

name                         old allocs/op  new allocs/op  delta
NormalizeWriteRawRequest-12      381k ± 0%      287k ± 1%  -24.70%  (p=0.000 n=50+50)
```

Together with just recently merged PRs we've got 125 K saving in allocations.

```
$ benchstat bench-very-old.txt bench-new.txt
name                         old time/op    new time/op    delta
NormalizeWriteRawRequest-12    43.9ms ± 7%    36.8ms ± 7%  -16.22%  (p=0.000 n=47+46)

name                         old alloc/op   new alloc/op   delta
NormalizeWriteRawRequest-12    53.5MB ± 1%    44.5MB ± 1%  -16.95%  (p=0.000 n=50+50)

name                         old allocs/op  new allocs/op  delta
NormalizeWriteRawRequest-12      412k ± 0%      287k ± 1%  -30.25%  (p=0.000 n=50+50)
```