@marselester 2023-05-16
<img width="2555" alt="Screenshot 2023-05-16 at 11 43 27" src="https://github.com/parca-dev/parca/assets/823099/35a7343e-0816-4422-adca-d880f1531f8f">

<details>

<summary>Parca logs</summary>

```
$ ./parca --debug-normalize-addresses=true
ooooooooo.
`888   `Y88.
 888   .d88'  .oooo.   oooo d8b  .ooooo.   .oooo.
 888ooo88P'  `P  )88b  `888""8P d88' `"Y8 `P  )88b
 888          .oP"888   888     888        .oP"888
 888         d8(  888   888     888   .o8 d8(  888
o888o        `Y888""8o d888b    `Y8bod8P' `Y888""8o



level=info name=parca ts=2023-05-16T15:42:29.122832471Z caller=factory.go:52 msg="loading bucket configuration"
level=info name=parca ts=2023-05-16T15:42:29.135311909Z caller=badger.go:54 msg="Set nextTxnTs to 0"
level=info name=parca ts=2023-05-16T15:42:29.146257111Z caller=server.go:93 msg="starting server" addr=:7070
level=error name=parca ts=2023-05-16T15:43:09.225809636Z caller=symbolizer.go:315 component=symbolizer msg="failed to remove liner file" err="remove /tmp/parca-symbolizer-3663392888: no such file or directory"
level=error name=parca ts=2023-05-16T15:43:49.190579342Z caller=symbolizer.go:315 component=symbolizer msg="failed to remove liner file" err="remove /tmp/parca-symbolizer-1907039591: no such file or directory"
```

</details>

<details>

<summary>Parca Agent logs</summary>

```
$ sudo ./dist/parca-agent --debug-normalize-addresses=false --remote-store-address=:7070 --remote-store-insecure
ooooooooo.                                                  .o.                                            .
`888   `Y88.                                               .888.                                         .o8
 888   .d88'  .oooo.   oooo d8b  .ooooo.   .oooo.         .8"888.      .oooooooo  .ooooo.  ooo. .oo.   .o888oo
 888ooo88P'  `P  )88b  `888""8P d88' `"Y8 `P  )88b       .8' `888.    888' `88b  d88' `88b `888P"Y88b    888
 888          .oP"888   888     888        .oP"888      .88ooo8888.   888   888  888ooo888  888   888    888
 888         d8(  888   888     888   .o8 d8(  888     .8'     `888.  `88bod8P'  888    .o  888   888    888 .
o888o        `Y888""8o d888b    `Y8bod8P' `Y888""8o   o88o     o8888o `8oooooo.  `Y8bod8P' o888o o888o   "888"
                                                                      d"     YD
                                                                      "Y88888P'

level=info name=parca-agent ts=2023-05-16T15:42:33.271743087Z caller=main.go:292 msg="maxprocs: Leaving GOMAXPROCS=2: CPU quota undefined"
level=info name=parca-agent ts=2023-05-16T15:42:33.283839489Z caller=main.go:334 msg="eBPF is supported and enabled by the host kernel"
name=parca-agent ts=2023-05-16T15:42:33.284664289Z caller=main.go:396 msg=starting... node=ubuntu-jammy store=:7070
level=info name=parca-agent ts=2023-05-16T15:42:33.286910742Z caller=main.go:499 msg=rlimit cur=1048576 max=1048575
level=info name=parca-agent ts=2023-05-16T15:42:33.295685422Z caller=cpu.go:224 msg="Attempting to create unwind shards" count=50
```

</details>